+++ spec/www/spec/sqlite-version-test.js
@@ -107,7 +107,7 @@ var mytests = function() {
 
               // Check specific SQLCipher version on plugin
               if (isAndroid) // (android-database-sqlcipher)
-                expect(rs.rows.item(0).myResult).toBe('3.34.1');
+                expect(rs.rows.item(0).myResult).toBe('3.39.4');
               else // SQLCipher for iOS/macOS/...
                 expect(rs.rows.item(0).myResult).toBe('3.39.4');
 
@@ -138,7 +138,7 @@ var mytests = function() {
 
               // Check specific SQLCipher version on plugin
               if (isAndroid) // (android-database-sqlcipher)
-                expect(rs.rows.item(0).cipher_version).toBe('4.4.3 community');
+                expect(rs.rows.item(0).cipher_version).toBe('4.5.3 community');
               else // SQLCipher for iOS/macOS/...
                 expect(rs.rows.item(0).cipher_version).toBe('4.5.3 community');
 
@@ -155,8 +155,7 @@ var mytests = function() {
 
         it(suiteName + 'math functions', function(done) {
           if (isWebSql) pending('SKIP for (WebKit) Web SQL');
-          // if (!isWebSql && isAndroid && isImpl2) pending('SKIP for plugin on Android with androidDatabaseImplementation: 2');
-          if (!isWebSql && isAndroid) pending('SKIP for plugin on Android'); // TODO
+          if (!isWebSql && isAndroid && isImpl2) pending('SKIP for plugin on Android with androidDatabaseImplementation: 2');
 
           var db = openDatabase('check-math-function.db');